image

+++ controllers/sauce.js
@@ -1,26 +1,63 @@
 const Sauce = require('../models/Sauce');
+const fs = require('fs');
 
 exports.createSauce = (req, res, next) => {
-    delete req.body._id;
-    const sauce = new Sauce({
-      ...req.body
-    });
-    sauce.save()
-    .then(() => res.status(201).json({ message: 'Sauce enregistrée !'}))
-    .catch(error => res.status(400).json({message: 'objet non créé !'}));
-  };
+  const sauceObject = JSON.parse(req.body.sauce);
+  delete sauceObject._id;
+  delete sauceObject._userId;
+  const sauce = new Sauce({
+      ...sauceObject,
+      userId: req.auth.userId,
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  });
 
-  exports.modifySauce =  (req, res, next) => {
-    Sauce.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-    .then( () => res.status(200).json({message: 'objet modifié !'}))
-    .catch(error => res.status(400).json({error}));
-  };
+  sauce.save()
+  .then(() => res.status(201).json({ message: 'Sauce enregistrée !'}))
+  .catch(error => res.status(400).json({message: 'objet non créé !'}));
+};
+
+
+exports.modifySauce = (req, res, next) => {
+  const sauceObject = req.file ? {
+      ...JSON.parse(req.body.sauce),
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : { ...req.body };
+
+  delete saucegObject._userId;
+  Sauce.findOne({_id: req.params.id})
+      .then((sauce) => {
+          if (sauce.userId != req.auth.userId) {
+              res.status(401).json({ message : 'Not authorized'});
+          } else {
+              Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
+              .then(() => res.status(200).json({message : 'Objet modifié!'}))
+              .catch(error => res.status(401).json({ error }));
+          }
+      })
+      .catch((error) => {
+          res.status(400).json({ error });
+      });
+};
+
 
   exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({_id: req.params.id})
-    .then( sauce => res.status(200).json({message: 'objet supprimé !'}))
-    .catch(error => res.status(400).json({error}));
-  };
+    Sauce.findOne({ _id: req.params.id})
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
   exports.getOneSauce = (req,res,next) => {
     Sauce.findOne({_id: req.params.id})
@@ -30,6 +67,7 @@ exports.createSauce = (req, res, next) => {
 
   exports.getAllSauces =(req, res, next) => {
     Sauce.find()
-    .then(sauces => res.status(20).json(sauces))
+    .then(sauces => res.status(200).json(sauces))
     .catch(error => res.status(400).json({error}));
   };
+
